Added a proper .travis.ml that builds all.
Travis now builds both touist-translator and touist-gui.
I also modified the touist-gui/build.xml to fit the
.travis.ml file.

On each tag, both osx and linux releases are going to be pushed to the
Github releases.

I also modified the README.md and INSTALL.md to be more compelling and
easier to understand.