Remove $ from variables.
To do that, I just removed the first character of every Var lexem, because the lexem text value was previously "$variable".

I didn't find a way not to have the $ in the lexem text value directly using the jflex generator.

In the end I chose to simply call a java function to remove the first char...