Displays "Translating" and then "Solving".
I did that modification because I think the user should know what is happenning under the hood, most importantly when the solver process is hanging for a long time. Is it because of the solver because the model is complex? Is it because of the translator to SAT because the model is compact (= unfolding it to SAT clauses takes time).