dimacs.ml: fixed bug of 'Clause is always true/false'
It appears when bigor/and empty.
To fix that, I changed the cnf to dimacs transformation
for it to handle the top/bottom symbols.

It was caused by a bad cnf translation of bot and top symbols;
bot and top are generated when bigand or bigor does not produce
anything.

Also, fixed dimacs print; now prints clause by clause (printf)
instead of storing the whole thing into a string