renamed all Syntax.exp to Syntax.ast
Rationale: we should REALLY separate the three types of data we use:
- ast would be the tree corresponding to what the user has entered
- form would be the tree corresponding to the formula evaluated,
  meaning that all bigand, bigor, exact, sets, variables... that do
  not belong to a formula (which should be only 'and', =>, or, not...).
- cnf is the transformed type that ONLY contains or, and and not.

Even if cnf is subset of form and
        form is subset of ast
we should separate them!!!! That would REALLY make things clearer!

In the meantime, I decided that the biggest set of symbols, ast, should
be the name of this current spagetti of types.