Affiche "Translation" puis "Solving" lorsqu'on appuie sur Solve (#133)
* Displays "Translating" and then "Solving".

I did that modification because I think the user should know what is happenning under the hood, most importantly when the solver process is hanging for a long time. Is it because of the solver because the model is complex? Is it because of the translator to SAT because the model is compact (= unfolding it to SAT clauses takes time).

* Removed "violent" popup "you interrupted translation!"

It is kind of useless to tell the user he interrupted the translation: he chose to push the Stop button (or push a second time the Solve button). So I removed that popup message.

* changelog