touistc: (HUGE) type checking moved from parse time to eval time
* type checking wasn't at the right place

I had to redesign the parser.mly because in expressions
like
    $i - a
because $i is reduced using int -> var instead of
prop -> var (because the is a reduce/reduce conflict,
and int -> var is the 'default' reduction arbitrarily
chosen).

Worse: when writing
    let $i=1.0: ($i - $i) + 3.0
and even though we deliberately know that $i contains a
float, the expression
    ($i - $i)
is reduced as an int...

Solution: do not type-check during parse type; it is painful
to design a parser that actually works without reduce/reduce
conflicts.

* unit tests added

I also added unit tests; to run them, do `make test`.
To allow these tests, I had to remove the main function from
touistc.ml. All the function are now in frontend.ml
and the main function is in touistc.ml.

* Oasis and build system

When doing 'make', we now create a library called 'Touist'
in the files touist.cma and touist.cmi. They contain basically
Translation, Smt and Dimacs

* String.uppercase_ascii only exists as of 4.03.0

But my build system is using 4.02.3. So we should not change
the `String.uppercase` to `String.uppercase_ascii` for now,
even if ocaml >= 4.03.0 tells it is deprecated.

* Cleaned errors raised when location is available