added a language documentation generated by circleci (#247)
The language reference/manual is written using Madoko (.mdk file).
I chose mdk over latex because markdown is way easier to type than
latex. The drawback is that Madoko doesn't have that much attention
and support... nevertheless, madoko comes with a nice online editor,
which helps a lot.

Note that madoko is still in active development.

See docs/CONTRIBUTE-TO-DOCS.md for more information.