gui: added the 'Others' option in result view
It allows us to filter any valuation that is different from True or False.
It is only available in SMT or QBF modes because the SAT mode only has
True and False values.

For example, it allows the user to remove (in QBF) any '?' (undefined)
valuations to help with readability.