touist --sat --solve: fixed bug that turned some unsat problems satisfiable
This was because I did not correctly understand/read the doc on
Minisat.add_clause_a. Instead of skipping the 'solve' because it
returned 'false', I was doing a solve anyway.

The thing is that after Minisat.add_clause_a turns false, any attempt
to add a new clause will be worthless (because minisat will not add this
clause anyway), leading the user with an instance of the solver that
does not contain the formula he thought he was solving.

Note: to be able to debug that, I added the 'verbose' mode of minisat when
using the option --debug.

Solves #290