opam: use a generated version.ml instead of %%VERSION%%
This is because on github the tarball was having %%VERSION%%
inside instead of the 'jbuilder subst' transformed version.

Usually, I should first create a source tarball with jbuilder subst
and then upload it to github. But I cannot... Because github creates
it for me.

But 'jbuilder subst' is still interesting because it will
add 'version: "3.5.0-2-gab98aff' (for example) if you are in a git
repo, which is interesting when people are pinning the repo.
If they download a github source tarball, they will just get the
version already in touist.opam, e.g., 3.5.0.