travis: removed 'set -e' so that travis doesn't fail with no log/error
The error 'shell_session_update' (whatever the line) was benign. The
problem was rather because I was using 'set -e' inside the .travis.yml
(as opposed to in a dedicated script file launched from .travis.yml).
This option was turning the benign internal travis errors into failures
with no apparent reason and with a truncated log.

Two ways to fix that:
1) not use 'set -e' in .travis.yml and instead put the 'set -e' in a
   dedicated script,
2) if 'set -e' is used in .travis.yml, then do not forget to reset the
   option with 'set +e'