Add a ./touist wrapper file so that it is easier to call touist in dev
Before jbuilder/dune, I was using ocamlbuild and oasis. What was nice
with the previous build system is that it was creating a symlink
to the native and bytecode executables at the root of the repo,
so it was easy to run the program:

    ./touist.native --solve <<< "a and b"
    ./touist.byte --solve <<< "a and b"

Now, using jbuilder, no symlink is done. Apparently, it is on purpose:

    https://github.com/ocaml/dune/issues/152